Add EntityHandle(32byte) Add AssetHandle(128byte, UUID) Add yaml-cpp Submodule Add MetaFileSystem(.ddmeta) Add AssetImporter Add Asset Add AssetManager Architecture Add FileDialog Submodule Modify GraphicsResourceManager

+++ Daydream/Source/Daydream/Graphics/API/DirectX12/D3D12Framebuffer.cpp
@@ -21,8 +21,8 @@ namespace Daydream
 	{
 		device = _device;
 		auto desc = _swapchainImage->GetDesc();
-		width = desc.Width;
-		height = desc.Height;
+		width = (UInt32)desc.Width;
+		height = (UInt32)desc.Height;
 		renderPass = _renderPass;
 
 		D3D12_RESOURCE_DESC textureDesc = _swapchainImage->GetDesc();
